requirements.txt
aiogram>=3.0.0
langchain
openai
python-dotenv
pytest

